Fix some bugs and re arrange the file structures.
