+++ MeDIT/Others.py
